requirements-cpu.txt
rich
click
loguru
numpy
pandas
pyzmq
jinja2
fastapi
python-multipart
uvicorn
websockets
torch 
torchvision 
opencv-python 
sentence-transformers